Add new templates and views for basket functionality

+++ deepay/apps/basket/models.py
@@ -44,6 +44,9 @@ class Basket(DefaultModel):
     locked = models.BooleanField(default=False)
 
     def add(self, inventory, qty):
+        if self.locked:
+            raise Exception("Basket is locked")
+
         basket_object, created = BasketObject.objects.get_or_create(
             basket=self, inventory=inventory
         )
@@ -54,6 +57,9 @@ class Basket(DefaultModel):
         basket_object.save()
 
     def remove(self, inventory):
+        if self.locked:
+            raise Exception("Basket is locked")
+
         basket_object = BasketObject.objects.get(basket=self, inventory=inventory)
         basket_object.delete()
